Fixed a bug related to django4 foreign key lookups
Previous versions of django ran a foreign key lookup with a const (1)
parameter in the column list.  Django 4 changed that to a parameterized
value and pass the parameter value 1.  I have created a new token type to
hold the aliased paramaterized value.  SQLConstParameterizedIdentifier
is subclassed from SQLConstIdentifier, and the original functionality is
broken out into SQLConstIntIdentifier.  This should allow anything that
was using isinstanceof for SQLConstIdentifier to still work and allows
expansion to other non-int constants.
A test case has been added for the django 4 select statement.